Move the MUC list modal to its own file
